image_micro/requirements.txt
flask
pillow
numpy 
opencv-python